+++ packages/observability/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/observability [1.0.0-alpha.33](https://github.com/anolilab/lunora/compare/@lunora/observability@1.0.0-alpha.32...@lunora/observability@1.0.0-alpha.33) (2026-08-21)
+
+
+### Dependencies
+
+* **@lunora/shard-engine:** upgraded to 1.0.0-alpha.33
+
 ## @lunora/observability [1.0.0-alpha.32](https://github.com/anolilab/lunora/compare/%40lunora%2Fobservability%401.0.0-alpha.31...%40lunora%2Fobservability%401.0.0-alpha.32) (2026-08-19)